app/requirements.txt
waitress==2.1.2
flask==2.3.2
flask-sqlalchemy==3.0.5
mysqlclient==2.2.0
sqlalchemy-utils==0.41.1
jinja2==3.1.2
flask-restx
sqlalchemy
